Create CurrentDateTime interface
The methods to get the current date and time were removed from the
DefaultHandler into their own interface. This decision was made because
it is functionality that is seperate from the handler and also so we are
able to mock out the current date and time to test the handler.